Ensure image_data and axes are float64
Convert image_data to float64 and adjust x, y ranges.

+++ lauexplore/image/_plot.py
@@ -22,14 +22,15 @@ def plot(
     if isinstance(image_data, (str, Path)):
         title = title or str(image_data)
         image_data = read(image_data)
-    
+
+    image_data = image_data.astype(np.float64)
     image_data = np.flipud(image_data)
     
     ny, nx = image_data.shape
     
     customdata, hovertemplate = plots.base_hovermenu(nx, ny)
-    x = np.arange(nx)
-    y = np.arange(ny)
+    x = np.arange(nx, dtype=np.float64)
+    y = np.arange(ny, dtype=np.float64)
     
     image_plot = plots.base.heatmap(
         image_data, x, y,
